requirements.txt
pyspark
boto3
pyyaml
nltk
numpy
mlflow

# pip install -e .  --break-system-packages